reduce step size in PowerLayer gradient checks: fix #1252
The gradient checker fails on certain elements of the PowerLayer checks,
but only 1-3 sometimes fail out of the 120 elements tested. This is not
due to any numerical issue in the PowerLayer, but the distribution of
the random inputs for the checks.

boost 1.56 switched the normal distribution RNG engine from Box-Muller
to Ziggurat.