[pycaffe] take pre-processing from Net and give to Transformer
The pyreformation #1703 aligned pycaffe Net with the real Caffe Net.
This broke the pre-processing helpers that had been bolted on to
`caffe.Net`.

These responsibilities are gathered in `caffe.io.Transformer` instead.
This is only an intermediate step in a real solution to pycaffe IO that
should make use of the same data pipeline as the rest of the framework.

See #1245 for further thoughts on input processing.