[pycaffe] switch examples to Transformer
This is a quick translation of the examples to the caffe.io.Transformer
interface. The results are not strictly identical to the earlier
implementation! The models now use a mean pixel instead of a mean image
for simplicity. The output classifications and detections are preserved
but scores may differ.

Note: the examples will be rewritten to make use of `caffe.Net` alone
since it is the true interface, but not yet.