Add a new script for the ImageNet data extraction
Provide a script that automates and simplifies the process of extracting
the .tar files which contain the ILSVRC2012 images (train and val). The
script provides comments on the stage of the process. It might be useful
for other purposes as well (extracting other .tar files of images) with
minor changes.