Add a new script to extract data for the ImageNet
Provide a script that automates and simplifies the process of extracting
the .tar files which contain the ILSVRC2012 images (train and val). The
script provides comments on the stage of the process. This script might be
useful for other purposes as well (extracting other .tar files of images)
with minor changes.

Also updated the readme.md file with information for the usage of the new
script.