Further refine auto-tuning for spatial convolution.
The type 4's 1,1,1 block is the slowest basic kernel which
we could use as a fallback solution when other kernels'
failed. No need to add it as an candidate.

And for two large block size , which means (x*y*z > 32),
we will not get better performance as limited registers.
So we limit x*y*z to less or equal to 32.

We could reduce more than 80% of the tuning time to get
the same performance.

Signed-off-by: Zhigang Gong <zhigang.gong@linux.intel.com>