Create "in_place" versions for common caffe_<math op> vector functions
Caffe implements a series of functions that operate on vectors, including
caffe_add, caffe_mul and caffe_div (there are more, but these are the
ones that seem to be used the most based on profiling).

These functions implement a vector  y = a <op> b  operation.

Now there is a slight inefficiency with this;
the compiler will nicely auto-vectorize these functions for the general
case, but because it can't prove that a, b and y are not overlapping aliases of eachother,
so it will also generate a non-SIMD version to fall back to.
(with aliases, the autovectorization in compilers, especially gcc, breaks
down so they fall back to the slow path).

this is all great, but there is a common pattern in the caffe code
that really wants to do

a = a <op> b

or in other words, an in-place add / mul / div.

This common pattern will ensure that "y" and "a" will alias, and this
then ends up executing the fallback, non-vectorized code, which is 4x or 8x
slower than the vectorized version.

This patch provides a solution to this problem by introducing "in_place" versions
of these 3 common operations; the next patch will introduce the users of these
new functions.

In the CPU profiles, the caffe_add/mul/div end up consuming several percent
CPU time (cpu-only configuration) and with this patch, that reduces drastically
giving a few percent (2 to 3% depending on hardwar) performance improvement on
the alexnet benchmark.