Refine spatial convolution kernel to support varying sizes.
To make each tuning for one single const image size is not a good solution.
In real application, users may want to process many different image sizes.
If we need to re-tune for each different size which even very close to each
other, it will be a big issue. So this patch is to relax the size limitation.

But this change introduce a little bit higher register pressure which will
cause performance regression which I will fix shortly.

Signed-off-by: Zhigang Gong <zhigang.gong@linux.intel.com>