Fix one incorrect constant value type.
We should always use Dtype constant value by default. Otherwise
it may cause compilation error for the platform doesn't support
double type. And even if the platform does support double, when
we build float type kernel, it will promote some of the routines
to double type which is not efficient.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>