Promote the URLs of individual layer doc pages.
Move the individual layer pages up one level, to make it easier to
promote the layers documentation in the future. This will prevent
breaking links from external websites, as long as we don't decide they
should actually be located somewhere else.

The layers.md is being left in tutorial for now, because it already has
a lot of external links.

Also, fix broken links in layer documentation.