Enable OCL version of hdf5_data_layer to fix a race condition
If we simply use the cpu code path to copy the data, we will introduce
one race condition between the GPU queue and the CPU. The scenario is:
when we call it in an iteration loop. The data blob is in a zero-copy
blob, and the first pass may be still blocking on the GPU side. The
second pass will modify the data blob on CPU side before the data is
accessed at the first pass on GPU side.

We can simply add a synchronization point between the two iterations, but that
is not a good fix as we force the GPU queue to flush and wait it to finish.
The best way is to do the copy on the GPU side and in the same queue. Thus we
don't need to worry about this race condition any more and without any interfere
the GPU queue.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>