libdnn: add spatial convolution implmentation
A new class in libdnn LibDNNConvSpatial which is porting
from clcaffe's ConvolutionLayerSpatial class but with following
changes:
 - Use basic kernel as benchmark kernel
 - Add LibDNNConvConfig::phase_test to indicate using in TEST phase

Also, the kernel source from conv_layer_spatial.cl is refined to
reduce reduntance and embedded into host code, so that it has no
dependence on Caffe components.

Note that the tuner mechanism is different from LibDNNTuner in
this version. It will cache the tuned parameters on local storage.
See "Auto-tuning and kernel cache mechanism" in
https://github.com/01org/caffe/wiki/clCaffe for details.