Avoid a potential compiler issue for FP16 convolution kernel.
The sub group shuffle extension has a restrication which it requires
all channels are active when we use this builtin functions. If we
do need to use those functions in if condition and need to access
cross work-item value, we need to apply some tricks to avoid the
potential compiler issue.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>