Fix bug in data_reader in which the 2nd data read on same dataset will bring system hang.
The original intention of using global varible bodies_ in data_read is wanting to reuse existing internalthread if the data source is same, but this idea doesn't work at all because the thread is waiting for old semaphore rather than the new one. this thread will always in waiting status after the first read.