Work around another driver bug.
We found for some version of CPU OCL device drivers, the cl_intel_subgroups is
predefined but it doesn't support it. So if it is the case, we need to undefine
cl_intel_subgroup extension.

Another issue is that for some Intel driver and platforms, the cl_intel_subgroups
is supported but cl_intel_subgroups_short is not. For these cases, the fp16 version
is broken and we just abort without compiling the unsupported kernels.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>