Fix a sub group related issue in Intel spatial convolution kernel.
The sub group extension requires to make sure all work items reach or
don't reach those APIs together. This is not the case for the output
phase as we will only write the result out when the position is valid.

This is the root cause why we need an extra work around code to prevent
the compiler to do wrong optimization.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>