Fix a weights related bug in intel spatial engine.
For some applications may change the input images frquently
and may use different kernel for the same layer with different
image size or different batch size. One corner case is that
the two kernels are the same type but with different SIMD
width. We still need to re-generate the weights for that case.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>