Need regenerate weights When the data is changed externally for FC/CONV layers.
We do weights conversion in FC and CONV layers. In training phase, we will always
do the conversion for each forwarding. And for test phase, we assumed the weights
will not changed which is not true when the test net is used with a training instance
and the training instance may change the weights data periodically.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>