remove the checks in convert_to_prv() as the same check exists in the following create_reorder_to_prv() invocation.
And this removal is also used to solve weight data/diff state mismatch issue, that's the case that sometimes the data is in SYNCED_PRV state but diff is in HEAD_AT_CPU state. The blob update() logic will update the blob state to HEAD_AT_CPU. if we don't remove the assertion it will hang at convert_to_prv() invoked at sync_before_read().